requirements.txt
PyCryptodome>=3.9.0 